+++ components/Cheese/Cheese.css
@@ -44,7 +44,7 @@
   width: 100%;
   height: 100%;
   text-align: center;
-  text-shadow: 0px 1px 2px #000000;
+  /*text-shadow: 0px 1px 2px #000000;*/
 }
 
 .cheese {
