+++ components/Header/Header.jsx
@@ -61,7 +61,9 @@ export default ({
       <ToolbarGroup>
         {navbarPages.map((page) => (
           <div key={page.name} className={styles.tabContainer}>
-            {page.external ? <a href={page.path} className={styles.tab}>{page.name}</a> : <Link to={page.path} className={styles.tab}>{page.name}</Link>}
+            {page.external ?
+              <a href={page.path} className={styles.tab}>{page.name}</a> :
+              <Link to={page.path} className={styles.tab}>{page.name}</Link>}
           </div>
         ))}
       </ToolbarGroup>
