+++ components/Heroes/index.js
@@ -1,11 +1,9 @@
 import Heroes from './Heroes';
-import HeroesIndex from './HeroesIndex';
 import Benchmark from './Benchmark';
 import Ranking from './Ranking';
 
 export {
   Heroes,
-  HeroesIndex,
   Benchmark,
   Ranking,
 };
