+++ index.js
@@ -14,7 +14,7 @@ import Player from './components/Player';
 import Home from './components/Home';
 import Search from './components/Search';
 import Explorer from './components/Explorer';
-import { Heroes, HeroesIndex, Ranking, Benchmark } from './components/Heroes';
+import { Heroes, Ranking, Benchmark } from './components/Heroes';
 import Request from './components/Request';
 import Distributions from './components/Distributions';
 import injectTapEventPlugin from 'react-tap-event-plugin';
@@ -66,11 +66,9 @@ render(
           </Route>
         </Route>
         <Route path="explorer" component={Explorer} />
-        <Route path="heroes" component={Heroes}>
-          <IndexRoute component={HeroesIndex} />
-          <Route path="rankings/:hero_id" component={Ranking} />
-          <Route path="benchmarks/:hero_id" component={Benchmark} />
-        </Route>
+        <Route path="heroes" component={Heroes} />
+        <Route path="rankings/:hero_id" component={Ranking} />
+        <Route path="benchmarks/:hero_id" component={Benchmark} />
         <Route path="request" component={Request} />
         <Route path="distributions" component={Distributions} />
         <Route path="carry" />
