+++ lang/en.json
@@ -1,0 +1,18 @@
+{
+	"gold_per_min": "Gold per minute",
+	"percentile": "Percentile",
+	"xp_per_min": "Experience per minute",
+	"last_hits_per_min": "Last hits per minute",
+	"kills_per_min": "Kills per minute",
+	"hero_damage_per_min": "Hero damage per minute",
+	"hero_healing_per_min": "Hero healing per minute",
+	"tower_damage_per_min": "Tower damage per minute",
+	"abbr_gold_per_min": "GPM",
+	"abbr_percentile": "%",
+	"abbr_xp_per_min": "XPM",
+	"abbr_last_hits_per_min": "LHM",
+	"abbr_kills_per_min": "KPM",
+	"abbr_hero_damage_per_min": "HDM",
+	"abbr_hero_healing_per_min": "HHM",
+	"abbr_tower_damage_per_min": "TDM"
+}
