--- language/en/index.js
@@ -1,7 +1,0 @@
-import tooltips from './tooltips.json';
-import abbrv from './abbreviation.json';
-
-export {
-  tooltips,
-  abbrv,
-};
