--- language/en/tooltips.json
@@ -1,10 +1,0 @@
-{
-	"gold_per_min": "Gold per minute",
-	"percentile": "Percentile",
-	"xp_per_min": "Experience per minute",
-	"last_hits_per_min": "Last hits per minute",
-	"kills_per_min": "Kills per minute",
-	"hero_damage_per_min": "Hero damage per minute",
-	"hero_healing_per_min": "Hero healing per minute",
-	"tower_damage_per_min": "Tower damage per minute"
-}
